+++ EatPic-iOS/Sources/Screens/Home/Views/Notification/NotificationView.swift
@@ -45,6 +45,78 @@ struct NotificationView: View {
     }
 }
 
+private struct NotificationItemView: View {
+    let notification: NotificationModel
+    let onTap: () -> Void
+    let onFollowTap: () -> Void
+
+    var body: some View {
+        Button(action: onTap) {
+            switch notification.type {
+            case .like, .comment:
+                // LikeNotiView 내용 직접 작성
+                HStack {
+                    ProfileImageView(size: 47)
+                    Spacer().frame(width: 16)
+                    (
+                        Text(
+                            "\(notification.friendNickname) 님이 회원님의 식사 기록에 "
+                            + (notification.type == .like ? "좋아요를 눌렀습니다." : "댓글을 남겼습니다.")
+                        )
+                            .foregroundColor(.gray080)
+                            .font(.dsFootnote)
+                        +
+                        Text("\(notification.notiTime) 전")
+                            .foregroundColor(.gray060)
+                            .font(.dsFootnote)
+                    )
+                    Spacer()
+                    // 좋아요, 댓글 알림은 항상 이미지가 있음
+                    if let postImageName = notification.postImageName {
+                        Image(postImageName)
+                            .resizable()
+                            .frame(width: 47, height: 47)
+                            .cornerRadius(4)
+                    }
+                }
+                .padding(.vertical, 12)
+                .padding(.horizontal, 20)
+                .frame(maxWidth: .infinity)
+                .background(notification.state.backgroundColor)
+
+            case .follow:
+                // FollowNotiView 내용 직접 작성
+                HStack {
+                    ProfileImageView(size: 47)
+                    Spacer().frame(width: 16)
+                    VStack(alignment: .leading) {
+                        Text("\(notification.friendNickname) 님이 회원님을 팔로우합니다.")
+                            .foregroundColor(.gray080)
+                            .font(.dsFootnote)
+                        Text("\(notification.notiTime) 전")
+                            .foregroundColor(.gray060)
+                            .font(.dsFootnote)
+                    }
+                    Spacer()
+                    Button(action: onFollowTap) {
+                        Text("팔로잉")
+                            .foregroundColor(notification.state.followButtonTextColor)
+                            .font(.dsBold15)
+                            .frame(width: 64, height: 28)
+                    }
+                    .background(notification.state.followButtonColor)
+                    .cornerRadius(5)
+                }
+                .padding(.vertical, 12)
+                .padding(.horizontal, 20)
+                .frame(maxWidth: .infinity)
+                .background(notification.state.backgroundColor)
+            }
+        }
+        .buttonStyle(.plain)
+    }
+}
+
 #Preview {
     NotificationView()
 }
